test(refactor): 테스트 패키지 구조를 본 코드와 동일하게 재구성
