+++ src/test/java/lotto/domain/WinningLottoNumberTest.java
@@ -1,6 +1,5 @@
-package lotto;
+package lotto.domain;
 
-import lotto.domain.WinningLottoNumber;
 import lotto.global.error.ErrorCode;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
