+++ src/test/java/lotto/util/ValidatorTest.java
@@ -1,13 +1,11 @@
-package lotto;
+package lotto.util;
 
 import lotto.global.error.ErrorCode;
-import lotto.util.Validator;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
 
-import java.util.ArrayList;
 import java.util.List;
 
 import static org.assertj.core.api.Assertions.*;
